1.Modified the sample code to cope with trampoline

+++ Projects/N76E003_BSP_Keil_C51_V1.0.6/Common/Common.c
@@ -73,7 +73,7 @@ UINT8 Receive_Data_From_UART0(void)
     return (c);
 }
 
-void Send_Data_To_UART0 (UINT8 c)
+void Send_Data_To_UART0 (UINT16 c)
 {
     TI = 0;
     SBUF = c;
